# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/exchange-codes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exchange Codes"
                ],
                "summary": "List exchange codes",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExchangeCodeListResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponseDoc"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exchange Codes"
                ],
                "summary": "Create exchange code",
                "parameters": [
                    {
                        "description": "Exchange code",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ExchangeCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.OKResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponseDoc"
                        }
                    }
                }
            }
        },
        "/api/v1/exchange-codes/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exchange Codes"
                ],
                "summary": "Get exchange code",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Exchange code ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ExchangeCodeSummaryResponseDoc"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponseDoc"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exchange Codes"
                ],
                "summary": "Update exchange code",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Exchange code ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Exchange code",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ExchangeCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.OKResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponseDoc"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponseDoc"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exchange Codes"
                ],
                "summary": "Delete exchange code",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Exchange code ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.OKResponseDoc"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponseDoc"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponseDoc"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.APIErrorResponseDoc"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.APIErrorResponseDoc": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/types.APIError"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "handlers.ExchangeCodeListResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.ExchangeCodeListResponse"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "handlers.ExchangeCodeSummaryResponseDoc": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/types.ExchangeCodeSummary"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
...
        "handlers.OKResponseDoc": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                }
            }
        },
...
        "types.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "types.ExchangeCodeListResponse": {
            "type": "object",
            "properties": {
                "codes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ExchangeCodeSummary"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/types.PaginationMeta"
                }
            }
        },
        "types.ExchangeCodeRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "quota": {
                    "type": "integer"
                },
                "rewards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ExchangeReward"
                    }
                }
            }
        },
        "types.ExchangeCodeSummary": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "platform": {
                    "type": "string"
                },
                "quota": {
                    "type": "integer"
                },
                "rewards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ExchangeReward"
                    }
                }
            }
        },
        "types.ExchangeReward": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
...
        "types.PaginationMeta": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}